requirements.txt
appdirs==1.4.4
black>=23.0.0
certifi==2025.10.5
charset-normalizer==3.4.4
feedparser==6.0.11
idna==3.11
ifaddr==0.2.0
lxml==6.0.2
mypy>=1.7.0
pillow>=10.0.0
pytest==8.3.3
pytest-mock==3.14.0
PyYAML==6.0.2
requests==2.32.5
requests-mock==1.12.1
soco==0.30.12
streamdeck==0.9.8
types-requests>=2.31.0
urllib3==2.5.0
xmltodict==1.0.2
